refactor: refactor highlight folder structure & more
