+++ src/scripts/highlighter/utils/createHighlight/createRangesClaude.ts
